jrgb:
I noticed that when you modify a model link in the alts toolkit that
it does not retain the pipe flow type and link type values.  For
example, when attempting to modify an existing storm pipe, the pipe
flow /link types default back to "C".  I'm not sure if this was
intentional but it can cause issues when one attempts to modify
something other than a combined sewer pipe.  Thx, Ruben.

jhb:
Updated link toolpad so default link/flow types are read from the mdl_link,
rather than defaulting to "C". Change tested and application updated.